Disable tests that were broken by the change to deaggregation.
All of these tests are directly or indirectly broken by the `simplify-arm` and `simplify-branch`
optimizations no longer firing. Unfortunately, this is a non-trivial thing to fix. Furthermore,
there is an existing open PR that completely re-writes the `simplify-arm` optimization. As such,
it doesn't make sense to try and fix the soon-to-be-replaced pass.