Re-enable some gdb tests
Bug #47163 points out that some debug tests were disabled.  This patch
re-enables many of them for gdb (I don't know enough about recent
versions of lldb to make this change there).  One test that only works
for a relatively new gdb, due to the use of DW_ATE_UTF, is moved to a
version-specific test.  In a couple of spot, I simplified the check
expression where I thought that it made sense to do so -- sometimes it
isn't very worthwhile to match exactly what is emitted, particularly
where this has changed over time.  I tested this with both a
DW_ATE_UTF-capable gdb, and an older version.