Add support for `sysroot` --extern flag
This adds `sysroot` to the set of extern flags:
`--extern sysroot:core=/path/to/core/libcore.rlib`.

The primary effect of this flag is to suppress `unused-crate-dependencies`
warnings relating to the crate.

It's used for build environments where it's desirable for the build
system to explicitly specify dependencies on sysroot crates rather than
letting rustc find them implicitly. In such cases, these dependencies
are likely to be added unconditionally, so there's no way that a user
could act on an unused dependency warning.