`impl<T: AsRef<Path>> Div<T> for {Path, PathBuf}`
This was suggested by @mgattozzi in https://twitter.com/mgattozzi/status/1511150559453229059.
Don't let your memes be dreams!

The serious motivation for this change is that it's currently annoying to do this in a
cross-platform way without unnecessary allocations. The possible options are:
1. (least common): `["x", "y", "z"].into_iter().collect::<PathBuf>().` This works, but is non-trivial to discover and very rare in practice.
2. (more common) `Path::new("x").join("y").join("z")`. This works, but does an unnecessary allocation for each intermediate path.
3. (most common) `Path::new("x/y/z")`. Not cross-platform.

This PR introduces a fourth option, `Path::new("x") / "y" / "z"`, which
is easier to type than both options 1 and 2, while still being cross-platform
and avoiding allocations.