Split `AttrArgs` from `MacArgs`.
`MacArgs` is confusing. It has three variants. It's used in two
different cases.
- Attributes, where all three variants are used.
- Function-like macro calls, where only one variant is used.

This commit separates these two cases. `AttrArgs` keeps much of the
functionality of the old `MacArgs`. The new `MacArgs` is a greatly
streamlined version of the old `MacArgs`. There is a small amount of
code duplication, but I think this change improves readability. It also
lets a few unreachable code paths (e.g. `unwrap`/`expect` calls) be
removed.

The commit also renames `MacDelimeter` as `AttrMacDelimeter`, because
it's used for both `AttrArgs` and `MacArgs`. For this type, it makes
more sense to share a single type between the two cases because all
variants and methods are used in both cases.