Remove SeqCst usage from Arc
Prior to this commit, there were multiple loads from the weak and strong
counts that used SeqCst ordering. For plain loads, SeqCst behaves like
Acquire with the additional guarantee that all *sequentially consistent*
operations are observed in the same order by all threads.

However, this additional guarantee does not apply here because we never do
any store or load-with-store operations that use SeqCst ordering on either
of these counts.

This commit relaxes the ordering to Acquire to better reflect the
requirements and guarantees of this code to the reader, compiler, and
hardware.