Use unchecked_add in checked_add to make it easier to optimize
See https://github.com/llvm/llvm-project/issues/53978
Using overflowing_add seems to make it harder for LLVM to optimize
since it returns the wrapped value, which we never use and we rely on
it noticing that we only use it after checking !overflow.
This also changes checked_sub and mul.