Include the rustc_apfloat crate in COPYRIGHT
The rustc_apfloat crate is a port of LLVM's APFloat code and was
introduced in https://github.com/rust-lang/rust/pull/43554. At that
time, LLVM was licensed under the UIUC license.

Since that time, contributors have made changes to the code without
being advised of the UIUC licensing resulting in the code base being a
mixture of UIUC licensed code and Apache 2/MIT licensed code.

Going forward, the compiler team will be asking contributors to license
their contributions under all of Apache 2/MIT/UIUC licenses so that,
eventually, either we can relicense the crate as simply UIUC licensed or
create a new port that is Apache 2/MIT licensed which can use the
changes contributed on top of the UIUC base.